Fixed issues related to network connection failures and logging in
- Fixed an issue where failing to connect to a remote server would
cause an unexpected error. This error was not handled properly, but now
activities can react to it

- Fixed an issue where failing to connect to a remote server when a
user is already logged in did not set up sign up/login buttons. This
prevented users from trying to log in/sign up if a server was down

- Added a very simple list view to the ManageTracksActivity layout to
prepare for adding functionality to allow users to view existing
track/playlist data